Add a basic firefox extension that does not communicate with switcher yet
The plan is to us it as a tab lister and switcher that receives commands
from the main switcher.

When the user starts firefox, the extension starts, and it starts a
'native application' (as called in firefox extension wiki) to be used
as a proxy for requests from the main switcher.
The API proxy (in addition to the extension and the main switcher)
is needed since the extension only communicate with firefox' native
messaging feature) that cannot send messages to existing external
applications, but only to start external applications and talk to them.
And, since the main switcher is running in the background (from the
moment of the first Ctrl+Alt+w), it cannot be started by the firefox
extension.